Added chat adding and updating

+++ migrations/20171126121307_create_users_table.js
@@ -1,7 +1,7 @@
 
 exports.up = function (knex) {
   return knex.schema.createTable('users', function (table) {
-    table.integer('id').unique()
+    table.bigInteger('id').unique()
     table.string('name', 255).unique()
     table.string('theme', 40)
     table.timestamps(['created_at', 'updated_at'])
